separate state de/encoding into a util to avoid circular dependencies
